+++ src/base_dumper/common/dump_format.py
@@ -4,5 +4,5 @@ from enum import Enum
 class DumpFormat(Enum):
     """Enum for dump type format."""
 
-    binary = 0
-    csv = 1
+    BINARY = 0
+    CSV = 1